requirements-dev.txt
-r requirements.txt
pytest-django==4.11.1
model-bakery==1.20.5